test locale: Delete commented-out test
This was supposed to have been commented out temporarily, but I'm not
aware of a way in which to create an invalid JSString that can't be
converted to UTF-8, at least not from Javascript. Probably the '\ud800'
triggered a bug in an earlier version of SpiderMonkey, but I don't think
this test can sensibly pass anymore.

https://bugzilla.gnome.org/show_bug.cgi?id=775444